+++ gradle.properties
@@ -1,2 +1,1 @@
 version=5.0.0.BUILD-SNAPSHOT
-org.gradle.daemon=true
