+++ spring-integration-sftp/src/test/java/org/springframework/integration/sftp/config/InboundChannelAdapterParserCachingTests-context.xml
@@ -3,18 +3,11 @@
 		xmlns="http://www.springframework.org/schema/integration"
 		xmlns:beans="http://www.springframework.org/schema/beans"
 		xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-		xmlns:p="http://www.springframework.org/schema/p"
-		xmlns:context="http://www.springframework.org/schema/context"
 		xmlns:util="http://www.springframework.org/schema/util"
-		xmlns:tool="http://www.springframework.org/schema/tool"
-		xmlns:lang="http://www.springframework.org/schema/lang"
 		xmlns:sftp="http://www.springframework.org/schema/integration/sftp"
 		xsi:schemaLocation="http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans.xsd
 						http://www.springframework.org/schema/integration http://www.springframework.org/schema/integration/spring-integration.xsd
-					 	http://www.springframework.org/schema/context http://www.springframework.org/schema/context/spring-context.xsd
                         http://www.springframework.org/schema/util http://www.springframework.org/schema/util/spring-util.xsd
-						http://www.springframework.org/schema/tool http://www.springframework.org/schema/tool/spring-tool.xsd
-						http://www.springframework.org/schema/lang http://www.springframework.org/schema/lang/spring-lang.xsd
 						http://www.springframework.org/schema/integration/sftp http://www.springframework.org/schema/integration/sftp/spring-integration-sftp.xsd">
 
 	<channel id="requestChannel">
@@ -24,7 +17,7 @@
 	<beans:bean id="sftpSessionFactory" class="org.springframework.integration.sftp.session.DefaultSftpSessionFactory">
 		<beans:property name="host" value="loclahost"/>
 		<beans:property name="knownHosts" value="local, foo.com, bar.foo"/>
-		<beans:property name="privateKey" value="classpath:org/springframework/integration/sftp/config/sftpTest"/>
+		<beans:property name="privateKey" value="classpath:org/springframework/integration/sftp/config/sftp_rsa"/>
 		<beans:property name="privateKeyPassphrase" value="ghj"/>
 		<beans:property name="password" value="hello"/>
 		<beans:property name="port" value="2222"/>
@@ -41,20 +34,20 @@
 	</util:properties>
 
 	<sftp:inbound-channel-adapter id="cachingAdapter"
-			session-factory="csf"
-			channel="requestChannel"
-			remote-directory="/foo"
-			local-directory="file:local-test-dir"
-			auto-startup="false">
+								  session-factory="csf"
+								  channel="requestChannel"
+								  remote-directory="/foo"
+								  local-directory="file:local-test-dir"
+								  auto-startup="false">
 		<poller fixed-rate="1000"/>
 	</sftp:inbound-channel-adapter>
 
 	<sftp:inbound-channel-adapter id="nonCachingAdapter"
-			session-factory="sftpSessionFactory"
-			channel="requestChannel"
-			remote-directory="/foo"
-			local-directory="file:local-test-dir"
-			auto-startup="false">
+								  session-factory="sftpSessionFactory"
+								  channel="requestChannel"
+								  remote-directory="/foo"
+								  local-directory="file:local-test-dir"
+								  auto-startup="false">
 		<poller fixed-rate="1000"/>
 	</sftp:inbound-channel-adapter>
