+++ spring-integration-sftp/src/test/java/org/springframework/integration/sftp/config/InboundChannelAdapterParserTests-context.xml
@@ -23,7 +23,7 @@
 	<beans:bean id="sftpSessionFactory" class="org.springframework.integration.sftp.session.DefaultSftpSessionFactory">
 		<beans:property name="host" value="loclahost"/>
 		<beans:property name="knownHosts" value="local, foo.com, bar.foo"/>
-		<beans:property name="privateKey" value="classpath:org/springframework/integration/sftp/config/sftpTest"/>
+		<beans:property name="privateKey" value="classpath:org/springframework/integration/sftp/config/sftp_rsa"/>
 		<beans:property name="privateKeyPassphrase" value="ghj"/>
 		<beans:property name="password" value="hello"/>
 		<beans:property name="port" value="2222"/>
@@ -44,6 +44,7 @@
 								  local-filter="acceptAllFilter"
 								  max-fetch-size="42"
 								  delete-remote-files="${delete.remote.files}"
+								  auto-startup="false"
 								  preserve-timestamp="true">
 		<poller fixed-rate="1000">
 			<transactional synchronization-factory="syncFactory"/>
