+++ spring-integration-sftp/src/test/java/org/springframework/integration/sftp/config/SftpInboundAutostartup-context.xml
@@ -10,7 +10,7 @@
 	<bean id="sftpSessionFactory" class="org.springframework.integration.sftp.session.DefaultSftpSessionFactory">
 		<property name="host" value="loclahost"/>
 		<property name="knownHosts" value="local, foo.com, bar.foo"/>
-		<property name="privateKey" value="classpath:org/springframework/integration/sftp/config/sftpTest"/>
+		<property name="privateKey" value="classpath:org/springframework/integration/sftp/config/sftp_rsa"/>
 		<property name="privateKeyPassphrase" value="ghj"/>
 		<property name="password" value="hello"/>
 		<property name="port" value="2222"/>
